+++ internal/media/types.go
@@ -118,7 +118,7 @@ type AdditionalEmojiInfo struct {
 }
 
 // DataFunc represents a function used to retrieve the raw bytes of a piece of media.
-type DataFunc func(ctx context.Context) (reader io.Reader, fileSize int, err error)
+type DataFunc func(ctx context.Context) (reader io.Reader, fileSize int64, err error)
 
 // PostDataCallbackFunc represents a function executed after the DataFunc has been executed,
 // and the returned reader has been read. It can be used to clean up any remaining resources.
